Changes to MTAConfigGenerator needed to be moved by hand
